start to move tests to correct locations
